requirements.txt
python-telegram-bot>=20.7
openpyxl>=3.1.2
